database creation
It turns out I wasn't using ansible's stat command correctly. The registered
variable was always true, whether the file existed or not. The database
creation command needed to check with a .stat.exists. Since it always evaluated
to true the database was never created. Then when the webapp started up it
created the database, but without initializing the schema. And so we ended
up with an empty database.

Fixes #10

Thanks for your patience with this one @shawngraham and @ianmilligan1